Refactor to modern React best practices
- Resolved ESLint fast-refresh warning by extracting ThemeContext to a separate `ThemeContextCore.ts` file.
- Encapsulated mobile device detection logic into a reusable `useIsMobile.ts` hook, removing duplication across components.
- Refactored AddTaskModal and AddSubtaskModal to use standard HTML forms, enhancing accessibility by enabling native "Enter" key submission.
- Created summary files documenting changes to each modified file.
