+++ public/version.json
@@ -1,6 +1,6 @@
 {
   "version": "1.5.11",
-  "buildDate": "2026-04-04T00:58:58.994Z",
+  "buildDate": "2026-04-07T15:53:23.782Z",
   "requiredUpdate": true,
   "updateMessage": "Version 1.5.11 is now available with new features. Please refresh to update!"
 }
