+++ src/context/ThemeContextCore.ts
@@ -1,0 +1,18 @@
+// src/context/ThemeContextCore.ts
+import { createContext, useContext } from 'react';
+
+export type ThemeMode = 'light' | 'dark';
+
+export interface ThemeContextType {
+  themeMode: ThemeMode;
+  toggleTheme: () => void;
+}
+
+// Create the context with a default value
+export const ThemeContext = createContext<ThemeContextType>({
+  themeMode: 'light',
+  toggleTheme: () => {},
+});
+
+// Custom hook to use the theme context
+export const useTheme = () => useContext(ThemeContext);
