+++ summary_of_changes_src_components_DarkModeToggle_tsx.md
@@ -1,0 +1,3 @@
+# Summary of Changes: `src/components/DarkModeToggle.tsx`
+
+*   **Updated Imports**: Updated `useTheme` import to pull from `../context/ThemeContextCore` instead of `../context/ThemeContext`. This aligns with the refactored context structure designed to resolve React Fast Refresh warnings.
