+++ summary_of_changes_src_components_Modals_AddTaskModal_tsx.md
@@ -1,0 +1,5 @@
+# Summary of Changes: `src/components/Modals/AddTaskModal.tsx`
+
+*   **Refactored to use standard HTML `<form>`**: Wrapped the modal inputs in a `<form onSubmit={handleSubmit}>` tag.
+*   **Prevented Default Form Behavior**: Updated `handleSubmit` to accept the form event and call `e.preventDefault()`.
+*   **Accessibility and Usability**: Using an actual form tag enables users to press 'Enter' inside the modal inputs to submit the task, greatly improving keyboard navigation and adhering to accessibility standards. Changed button types appropriately (`type="button"` for Cancel, `type="submit"` for Add Task).
