+++ summary_of_changes_src_components_TaskNavigationView_tsx.md
@@ -1,0 +1,3 @@
+# Summary of Changes: `src/components/TaskNavigationView.tsx`
+
+*   **Replaced custom mobile detection logic with `useIsMobile` hook**: Imported and utilized the `useIsMobile` hook to simplify the component and eliminate duplicate code for detecting mobile environments. The explicit `useEffect` handling media queries was removed.
